// src/types/bindings/Creator.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CreatorPackage } from "./CreatorPackage";

export type Creator = { id: string, name: string, handle: string, avatar: string, discipline: string, sub_skills: Array<string>, city: string, locality: string, starting_at: number, rating: number, review_count: number, bio: string, verified: boolean, portfolio_urls: Array<string>, packages: Array<CreatorPackage>, };