// src/types/bindings/Booking.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Booking = { id: string, creator_id: string, creator_name: string, creator_avatar: string, client_name: string, pkg_name: string, date_time: string, status: string, price: number, deposit_amount: number, balance_amount: number, location_type: string, };