// src/types/bindings/CreateBookingPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreateBookingPayload = { creator_id: string, pkg_name: string, date_time: string, location_type: string, start_time: string | null, end_time: string | null, client_notes: string | null, };