// src/types/bindings/CreatorOnboardPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreatorOnboardPayload = { name: string, bio: string, discipline: string, sub_skills: Array<string>, years_exp: number, upi_id: string, instagram_handle: string, portfolio_urls: Array<string>, };